requirements.txt
streamlit
pandas==2.1.0
tensorflow==2.15.0
numpy==1.25.0
pillow
scikit-learn==1.6.1
matplotlib
plotly
plotly.express
